move function create_random_request to spyre_utils.py
